ai-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pillow==10.3.0
numpy==1.26.4
